# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Simple Scrumboard",
        "version": "0.1"
    },
    "paths": {
        "/api/boards": {
            "get": {
                "tags": [
                    "Boards"
                ],
                "summary": "Get All Boards",
                "operationId": "8544b25ebe792c9b47f2d926e2c8ebf8",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "examples": {
                            "string": {
                                "summary": "Set Parameter Application json",
                                "value": "Application/json"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/JsonBaseTransformer"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Board"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonBaseTransformer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/boards/{id}": {
            "get": {
                "tags": [
                    "Boards"
                ],
                "summary": "Get Board by ID",
                "operationId": "2a5da2c5db5012e8f8d32e941099bfa8",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "examples": {
                            "string": {
                                "summary": "Set Parameter Application json",
                                "value": "Application/json"
                            }
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "examples": {
                            "int": {
                                "summary": "ID value",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/JsonBaseSingleTransformer"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Board"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonBaseTransformer"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Board": {
                "title": "Board Schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Backlog"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2023-05-20T15:47:19.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2023-05-20T15:47:19.000000Z"
                    }
                },
                "type": "object"
            },
...
            "JsonBaseTransformer": {
                "title": "Json Base Transformer",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "code": {
                        "type": "integer"
                    },
                    "http_code": {
                        "type": "integer"
                    },
                    "locale": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "additional": {
                        "type": "string"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "first": {
                                    "type": "string"
                                },
                                "last": {
                                    "type": "string"
                                },
                                "prev": {
                                    "type": "string"
                                },
                                "next": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "meta": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "current_page": {
                                    "type": "integer"
                                },
                                "from": {
                                    "type": "integer"
                                },
                                "last_page": {
                                    "type": "string"
                                },
                                "path": {
                                    "type": "string"
                                },
                                "per_page": {
                                    "type": "string"
                                },
                                "to": {
                                    "type": "string"
                                },
                                "total": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "data": {
                        "description": "Array of returned data",
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            },
            "JsonBaseSingleTransformer": {
                "title": "Json Base Single Transformer",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "code": {
                        "type": "integer"
                    },
                    "http_code": {
                        "type": "integer"
                    },
                    "locale": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "additional": {
                        "type": "string"
                    },
                    "links": {
                        "type": "array",
                        "items": {}
                    },
                    "meta": {
                        "type": "array",
                        "items": {}
                    },
                    "data": {
                        "description": "Returned data"
                    }
                },
                "type": "object"
            }
        }
    }
}